Moves custom calendar to separate component
This change moves the custom calendar component from
`src/components/ui/calendar.tsx` to `src/components/calendar-custom.tsx`.
